# web/frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOMO Study Agent Web UI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/words": {
      "get": {
        "tags": [
          "words"
        ],
        "summary": "List Words",
        "description": "分页列出 ai_word_notes，支持搜索和筛选。",
        "operationId": "list_words_api_words_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "sync_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sync Status"
            }
          },
          {
            "name": "it_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "It Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WordsListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/words/{voc_id}": {
      "get": {
        "tags": [
          "words"
        ],
        "summary": "Get Word Detail",
        "description": "获取单个单词的完整笔记详情。",
        "operationId": "get_word_detail_api_words__voc_id__get",
        "parameters": [
          {
            "name": "voc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WordNoteDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "words"
        ],
        "summary": "Update Word Note",
        "description": "编辑单词笔记的 memory_aid 字段。",
        "operationId": "update_word_note_api_words__voc_id__put",
        "parameters": [
          {
            "name": "voc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/words/{voc_id}/iterations": {
      "get": {
        "tags": [
          "words"
        ],
        "summary": "Get Word Iterations",
        "description": "获取单词的迭代历史。",
        "operationId": "get_word_iterations_api_words__voc_id__iterations_get",
        "parameters": [
          {
            "name": "voc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_WordIterationsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "UNKNOWN"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiError"
      },
...
      "ApiResponse_WordIterationsResponse_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordIterationsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[WordIterationsResponse]"
      },
      "ApiResponse_WordNoteDetail_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordNoteDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[WordNoteDetail]"
      },
      "ApiResponse_WordsListResponse_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordsListResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[WordsListResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WordIteration": {
        "properties": {
          "voc_id": {
            "type": "string",
            "title": "Voc Id"
          },
          "iteration_type": {
            "type": "string",
            "title": "Iteration Type"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 0.0
          },
          "justification": {
            "type": "string",
            "title": "Justification",
            "default": ""
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "default": ""
          },
          "refined_content": {
            "type": "string",
            "title": "Refined Content",
            "default": ""
          },
          "raw_response": {
            "type": "string",
            "title": "Raw Response",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "voc_id",
          "iteration_type"
        ],
        "title": "WordIteration"
      },
      "WordIterationsResponse": {
        "properties": {
          "voc_id": {
            "type": "string",
            "title": "Voc Id"
          },
          "iterations": {
            "items": {
              "$ref": "#/components/schemas/WordIteration"
            },
            "type": "array",
            "title": "Iterations"
          }
        },
        "type": "object",
        "required": [
          "voc_id"
        ],
        "title": "WordIterationsResponse"
      },
      "WordNoteDetail": {
        "properties": {
          "voc_id": {
            "type": "string",
            "title": "Voc Id"
          },
          "spelling": {
            "type": "string",
            "title": "Spelling",
            "default": ""
          },
          "basic_meanings": {
            "type": "string",
            "title": "Basic Meanings",
            "default": ""
          },
          "memory_aid": {
            "type": "string",
            "title": "Memory Aid",
            "default": ""
          },
          "it_level": {
            "type": "integer",
            "title": "It Level",
            "default": 0
          },
          "sync_status": {
            "type": "integer",
            "title": "Sync Status",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "ielts_focus": {
            "type": "string",
            "title": "Ielts Focus",
            "default": ""
          },
          "collocations": {
            "type": "string",
            "title": "Collocations",
            "default": ""
          },
          "traps": {
            "type": "string",
            "title": "Traps",
            "default": ""
          },
          "synonyms": {
            "type": "string",
            "title": "Synonyms",
            "default": ""
          },
          "discrimination": {
            "type": "string",
            "title": "Discrimination",
            "default": ""
          },
          "example_sentences": {
            "type": "string",
            "title": "Example Sentences",
            "default": ""
          },
          "word_ratings": {
            "type": "string",
            "title": "Word Ratings",
            "default": ""
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "default": ""
          },
          "raw_full_text": {
            "type": "string",
            "title": "Raw Full Text",
            "default": ""
          },
          "it_history": {
            "type": "string",
            "title": "It History",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "voc_id"
        ],
        "title": "WordNoteDetail"
      },
      "WordNoteSummary": {
        "properties": {
          "voc_id": {
            "type": "string",
            "title": "Voc Id"
          },
          "spelling": {
            "type": "string",
            "title": "Spelling",
            "default": ""
          },
          "basic_meanings": {
            "type": "string",
            "title": "Basic Meanings",
            "default": ""
          },
          "memory_aid": {
            "type": "string",
            "title": "Memory Aid",
            "default": ""
          },
          "it_level": {
            "type": "integer",
            "title": "It Level",
            "default": 0
          },
          "sync_status": {
            "type": "integer",
            "title": "Sync Status",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "voc_id"
        ],
        "title": "WordNoteSummary"
      },
      "WordsListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "default": 50
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WordNoteSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "WordsListResponse"
      }
    }
  }
}